+++ csharp/src/Verifier.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Security.Cryptography;
 using System.Text;
+using System.Text.Json;
 using Jose;
 
 namespace TrueLayer.Signing
@@ -24,6 +25,35 @@ namespace TrueLayer.Signing
         public static Verifier VerifyWithPem(ReadOnlySpan<byte> publicKeyPem)
             => VerifyWithPem(Encoding.UTF8.GetString(publicKeyPem));
 
+        /// <summary>
+        /// Start building a `Tl-Signature` header verifier using public key JWKs JSON response data.
+        /// </summary>
+        /// <exception cref="SignatureException">Jwks is invalid</exception>
+        public static Verifier VerifyWithJwks(string jwksJson) => VerifyWithJwks(jwksJson.ToUtf8());
+
+        /// <summary>
+        /// Start building a `Tl-Signature` header verifier using public key JWKs JSON response data.
+        /// </summary>
+        /// <exception cref="SignatureException">Jwks is invalid</exception>
+        public static Verifier VerifyWithJwks(ReadOnlySpan<byte> jwksJson)
+        {
+            try
+            {
+                var jwks = JsonSerializer.Deserialize<Jwks>(jwksJson, new JsonSerializerOptions
+                {
+                    PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
+                });
+                // ecdsa fully setup later once we know the jwk kid
+                var verifier = VerifyWith(ECDsa.Create());
+                verifier.jwks = jwks ?? new Jwks();
+                return verifier;
+            }
+            catch (JsonException e)
+            {
+                throw new SignatureException("invalid jwks", e);
+            }
+        }
+
         /// <summary>Start building a `Tl-Signature` header verifier usinga a public key.</summary>
         public static Verifier VerifyWith(ECDsa publicKey) => new Verifier(publicKey);
 
@@ -31,7 +61,7 @@ namespace TrueLayer.Signing
         /// <exception cref="SignatureException">Signature is invalid</exception>
         public static string ExtractKid(string tlSignature)
         {
-            var kid = Jose.JWT.Headers(tlSignature)["kid"] as string;
+            var kid = Jose.JWT.Headers(tlSignature).GetString("kid");
             if (kid == null)
             {
                 throw new SignatureException("missing kid");
@@ -39,7 +69,25 @@ namespace TrueLayer.Signing
             return kid;
         }
 
+        /// <summary>
+        /// Extract jku (JSON Web Key URL) from unverified jws Tl-Signature.
+        /// Used in webhook signatures providing the public key jwk url.
+        /// </summary>
+        /// <exception cref="SignatureException">Signature is invalid</exception>
+        public static string ExtractJku(string tlSignature)
+        {
+            var jku = Jose.JWT.Headers(tlSignature).GetString("jku");
+            if (jku == null)
+            {
+                throw new SignatureException("missing jku");
+            }
+            return jku;
+        }
+
         private ECDsa key;
+        // Non-null when verifying using jwks data.
+        // This indicates we need to initialize `key` once we have the kid.
+        private Jwks? jwks;
         private string method = "";
         private string path = "";
         private Dictionary<string, byte[]> headers = new Dictionary<string, byte[]>(new HeaderNameComparer());
@@ -148,9 +196,17 @@ namespace TrueLayer.Signing
         {
             var jwsHeaders = SignatureException.Try(() => Jose.JWT.Headers(tlSignature));
 
+            if (jwks is Jwks jwkeys)
+            {
+                // initialize public key using jwks data
+                var kid = jwsHeaders.GetString("kid") ?? throw new SignatureException("missing kid");
+                FindAndImportJwk(jwkeys, kid);
+            }
+
             SignatureException.Ensure(jwsHeaders.GetString("alg") == "ES512", "unsupported jws alg");
             SignatureException.Ensure(jwsHeaders.GetString("tl_version") == "2", "unsupported jws tl_version");
-            SignatureException.Ensure(tlSignature.Contains(".."), "signature must have a detached payload");
+            var signatureParts = tlSignature.Split(".");
+            SignatureException.Ensure(signatureParts.Length >= 3, "invalid signature format");
 
             var signatureHeaderNames = (jwsHeaders.GetString("tl_headers") ?? "")
                 .Split(",")
@@ -164,9 +220,29 @@ namespace TrueLayer.Signing
             var signedHeaders = FilterOrderHeaders(signatureHeaderNames);
 
             var signingPayload = Util.BuildV2SigningPayload(method, path, signedHeaders, body);
-            var jws = tlSignature.Replace("..", $".{Base64Url.Encode(signingPayload)}.");
+            var jws = $"{signatureParts[0]}.{Base64Url.Encode(signingPayload)}.{signatureParts[2]}";
 
             SignatureException.Try(() => Jose.JWT.Decode(jws, key), "Invalid signature");
+        }
+
+        /// <summary>Find and import jwk into `key`</summary>
+        private void FindAndImportJwk(Jwks jwks, string kid)
+        {
+            var jwk = SignatureException.Try(() => jwks.Keys.First(key => key.Kid == kid), "no jwk found with kid");
+
+            SignatureException.Ensure(jwk.Kty == "EC", "unsupported jwk.kty");
+            SignatureException.Ensure(jwk.Alg == "EC", "unsupported jwk.alg");
+            SignatureException.Ensure(jwk.Crv == "P-521", "unsupported jwk.crv");
+
+            SignatureException.TryAction(() => key.ImportParameters(new ECParameters
+            {
+                Curve = ECCurve.CreateFromFriendlyName("secp521r1"),
+                Q = new ECPoint
+                {
+                    X = Base64Url.Decode(jwk.X),
+                    Y = Base64Url.Decode(jwk.Y),
+                }
+            }), "invalid jwk data");
         }
 
         /// <summary>Filter and order headers to match jws header `tl_headers`.</summary>