+++ csharp/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.1.4
+* Fix `Verifier` allowing non-detached jws signatures with trailing dots.
+
+## 0.1.3
+* Add `Verifier.ExtractJku` to extract `jku` jws header from webhook signatures.
+* Add `Verifier.VerifyWithJwks` to aid verifying webhook signatures.
+
 ## 0.1.2
 * Add `Verifier` support for signatures without headers.
 * Fix `Verifier` allowing non-detached jws signatures.